Convert types/atomic/ferguson/atomictest to a performance test
Compare the performance of fetchAdd for uint(32), uint(64), int(32), int(64),
and real(64) atomics. The types being compared is limited based on the largest
integer value that will fit in any given type. i.e. this test does 100000000
fetchAdds, but the smaller int/uint types and real(32) can't store an integer
that large (in the real(32) case, can't *exactly* store an integer that large.)

This is meant to track performance of atomic operations over time and to give a
comparison between the different atomic types, particularly between Integral
and Real atomics since there's no hardware support for floating point atomics.

Note that we should really add more atomic performance tests in the future for
other operations and comparisons to sync vars, but this was easy to do since I
just updated this test with #4511