Adjust callDestructors to resolve issue with global aliases
This change resolves a problem with the following program:

    module bar {
      var dom = {1..10, 1..10};
      var A : [dom] int;

      var R => A[1..4, 1..4]; // call_tmp = slice
      // slice freed here
    }

    use bar;

    proc main() {
      writeln("Setting A[1,1] to 1");
      A[1,1] = 1;
      writeln("R[1,1] is now     ", R[1,1]);

      for r in R do write(r);
      writeln();
    }

Without this change, the call_tmp storing A[1..4, 1..4] is destroyed at the end
of module bar's init function. With this change, that call_tmp is not destroyed
at all; instead the ownership of it is transferred to the variable R.

In order to accomplish this change:
 * created FLAG_NEW_ALIAS_FN and marked newAlias with it
 * added a transformation in callDestructors to handle this case.

This version can be seen as a workaround. In the future, I'd like to similarly
support code like

  ref R = A[1..4, 1..4];

Additionally, this change was compilicated by the way that function resolution
adds FLAG_INSERT_AUTO_COPY and by the fact that normalize occurs before
resolution. It would be preferable to be able to handle this case in normalize.