Added reduce= operator.
This operator represenst the accumulation operation for a reduce intent.

* Added reduce= token.

* Parser converts "x reduce= y" to: PRIM_REDUCE_ASSIGN(x,y)

* At start of normalization, it is lowered to the call:
globalOp.accumulateOntoState(x,y)

* Added a check that there are no PRIM_REDUCE_ASSIGN calls in the tree
after resolution. Really, there shouldn't be any after normalization.
We do not already have a loop over all gCallExprs in the post-normalization
check, and I did not want to add one.

* Added accumulateOntoState() to our standard ReduceScanOp subclasses.

* While there, I removed the default value for the char* 'op' argument of
buildAssignment() - because that default was unused and, more importantly,
there is an assertion against that default in the function itself.