Accept change in error output
(discussed with Vass)

For some reason, changes on my arrays branch (probably the tuple rework)
has caused this test to only fail with 1 error at a time when it used to
fail with all of them.

In the future, we'd like for sync/single to be storable in a tuple, so in
some ways this test is a stopgap anyway.