Clean up privitized freeing
The previous version of the code for freeing privatized classes wasn't
actually doing anything because _isPrivatized(this) was always returning
false. The new version adjusts _delete_dist to take in an argument which
comes from ChapelArray records.

This is a bit tricky because we can't currently privatize or delete
privatized versions of DefaultRectangular because of resolution ordering
issues (DefaultRectangular used to create Locales array).