Replace argument references - leftover from stealing the init from the type's default constructor - with references to fields.
Traverse the initExpr we saved from building the default constructor.  This
expression has undergone scopeResolution in the context of its current place in
the tree (as part of the default expression for an argument in the default
constructor).  When we copy it to reuse for the initialization of omitted
fields, we must update the references that will no longer be valid - namely,
any reference to other fields would be represented as references to the
arguments of the same name to the default constructor.

Doesn't handle references to fields on a parent type yet.