Add test of when default value for field's type has been changed
I was worried when implementing omitted initialization of fields that the
solution I was following was relying on the stored defaultValue of the type
which we had tried to move away from relying on in my _defaultOf work from a
couple of years ago.  So I created this test to verify the behavior was correct.
As it turns out, I was right to be worried - though the code path I was
imitating (initializeClass from functionResolution) somehow doesn't result in
ignoring the _defaultOf function for fields, probably due to some action further
in function resolution, taking just that part did cause the _defaultOf function
to be ignored.

I think I have a fix for this, but I wanted to save the test first just in case.