Extend comment of TODOs in handleInitializerRules
Remind myself that in the case where no init call is present in the body of the
initializer, we will need to insert a super.init() call with no arguments as
well as default initialization for all the fields (especially once I've removed
the call to the default constructor)

Saving this now because I realized it would disappear with my git stash playing.
Probably should go in independent of the rest of the changes on this branch.