Remove future on test that now passes and add it to test that now breaks
With the addition of initialization for omitted fields, phase1/dependence.chpl
now passes.  However, my search for field initialization does not take into
account field initialization that occurs within if statements - so in
phase1/hasIf.chpl, we think that the field is missing initialization when in
fact its initialization is handled within the if statement.  I will handle if
statements (both for .init() calls and field initialization) as a later step.