Renewed attempts at eliminating inner multiplies by param fields
The main change in approach here is to try and parameterize the
Block*Arr classes with noInnerMult as well, and to make their
'myElems' fields be declared not using a default rectangular array,
but by threading the noInnerMult value through a clone of
dom.buildArray() especially for default rectangular arrays in
order to avoid the potential type mismatch when a => operator is
applied to a reindexed or rank-changed array.

Disappointingly reindexing and rankchange don't work, apparently due
to a type mismatch in the => argument (in that, if I comment that
argument out, things compile, but of course don't work).  I'm
wondering whether the compiler is building up this argument's type in
a way that makes it incompatible (?).  I also wonder what would happen
if the noInnerMult were hard-coded to false everywhere in BlockDist
(or everywhere altogether, just to get things compiling?).