Use an atomic spin lock instead of a sync var for reductions
Native qthread sync vars (#4506) resulted in some performance regressions for
tests that use reductions (lulesh, NAS CG, Reductions Time, etc.)

The core problem is that native qthread sync vars are great for long-lived and
heavily-contended locks, but not for short-lived and minimally-contended ones.

That's actually true for sync vars in general, but it's especially true for
native qthread sync vars currently. This is because qthread sync vars use a
global hashtable where a new entry is added when a sync variable is emptied,
and removed when a sync var is filled and has no pending operations left. This
means that sync vars used as locks are repeatedly being added and removed from
the hashtable, which is expensive. For reductions each task did something like:

``` chapel
var lock: sync bool;  // create empty var -- results in hash entry creation

lock.writeEF();       // lock -- fills, no pending ops, results in hash entry removal

accumReduction();

lock.readFE()         // unlock  -- empties, results in hash entry creation

// lock out of scope  // sync destructor fills, results in hash entry removal
```

which resulted in twice as many hash table manipulations as we want. This makes
me think that we need to modify qthreads so that we can control when sync vars
are added/removed from the hashtable instead of having them be automatically
removed for full variables that have no pending operations, because that's
going to be suboptimal for a lot of cases.

Regardless, the actual accumulation for reductions shouldn't be using sync vars
in the first place. Accumulation should be a short-lived, minimally-contended
step because we usually just use dataParTasksPerLocale tasks per reduction,
which makes an atomic lock ideal.

This commit swaps  out the sync var lock/unlock for an exponential backoff
atomic spin lock, loosely based off the one I added to meteor with (#4300.) In
my local tests, this actually results in better performance than we were seeing
prior to native qthread sync vars.