Change default binary string format to data_toeof
Previously, the default binary string format was variable-byte-length
followed by data. However, that format was confusing because

  textchannel.write(mystring)

outputs only the string data, but

  binarychannel.write(mystring)

would output a length in addition to the string data.

Now that we have formatted I/O (with writef and readf), it is more
reasonable to make these two writes do the same thing. Someone wanting
strings that can be read back in should be setting a string style or
using formatted I/O.