Snapshot of attempt to parameterize all arrays
There are some hacks in here (search for "want") compared to what
I wanted to pursue, but this neither worked with or without the
hacks.  I think the issue is that we want/need the original array
to be one value of the param and the compiler-introduced alias to
be the other; or else the compiler introduced alias doesn't have
the type we want.  Switching over to master to learn more.