Print an explanation upon const error due to task/forall intents
In addition to printing
  file.chpl:NNN: error: illegal lvalue in assignment

or similar, also print
  file.chpl:MMM: note: The shadow variable 'XXX' is constant due to forall intents in this loop

or "... due to task intents in this cobegin or coforall statement" or
"... in this begin statement", as appropriate.

Updated several .good with these additional notes.

WHILE THERE, I removed .prediff that squashed line numbers for a couple
of tests. We added squashing in #298 and I felt OK with it back then.
Now I feel that it is somewhat important to ensure that the line numbers
are reported correctly in error messages. In particular, that the
parallel construct for the task/forall intents is pointed at correctly.
So I removed squashing and locked in the line numbers reported currently.
I spot-checked that these line numbers are correct.

IMPLEMENTATION NOTES

I added FLAG_CONST_DUE_TO_TASK_FORALL_INTENT to mark the variables
that became const due to task/forall intents.

WHILE THERE, I also ensured that the line number associated with
the BlockStmt that contains (and calls) individual task functions
for a 'coforall' is the line number for the 'cobegin' keyword (I think).
That way my "notes" can point to that line, rather than the line
where the particular task function is located.