Add a compiler optimization that removes the inner multiply for array indexing when possible. The basic idea of the optimization is that if the blk field is never modified, we know that blk(rank) will always be 1, so we can optimize away the last blk multiply in dsiAccess().
In order to determine if blk is modified, we explicitly mark all functions that
might modify it with a pragma. Then, if none of those functions are called we
know blk cannot be modified, so it's safe to remove the multiply. The functions
that modify blk are adjustBlkOffStrForNewDomain() (called for strided reindexing
and strided slicing) and rankChange().

Note that we have to be careful to look for the functions that modify blk after
all functions have been resolved, but before we do param folding.

This is kind of a lame optimization because if any array requires blk, all
arrays will suffer the performance penalty. i.e. the optimization just checks
if any function modifies block, it doesn't do any contextual or type based
analysis. Longer term we expect array views will provide a way to remove the
inner multiply in a cleaner and more principled manner, so this optimization is
really just a stopgap until that comes along, and it happens to get us the
performance for almost all the cases we currently care about.

This has a dramatic performance improvement for things like the PRKs, array
microbenchmarks, lcals, and some of the shootout benchmarks.