Rename canResolveMethod() -> canResolveTypeMethod() for type methods
Michael points out that not only is this clearer, but it opens the
way for being able to reflect on a call for which a value may not
occur in the future (by providing a version of canResolveMethod()
that takes a type rather than an object.

Also some additional minor wordsmithing.