Added reduce= operator.
This operator represents the accumulation operation for a reduce intent.
For example, when using third-party reduction operations,
the user does not need to be concerned with what the accumulation
operator is - just use "reduce=":

    var x ...;
    forall ... with (CoolReduction reduce x) {
      x reduce= some-computation;
    }

Extended our standard reductions to support reduce=
as per the current UDR interface - see below.

SHADOW VARIABLE STORES ACCUMULATION STATE

In our group discussion on 9/06/2016, we discussed the semantics
of the outer vs. shadow variable that is subject to a reduce intent.
Reminder: in the above example, "x" outside the loop is the "outer"
variable and "x" inside the loop is the corresponding "shadow" variable.

We concluded:

* The outer variable has the type of - and, after the loop, stores
  the value of - the reduction result.

* The shadow variable is task-private. (This has been this way
  since reduce intents were introduced.)

* The shadow variable stores the current accumulation state
  within its task. It could be examined and operated upon,
  besides being on the l.h.s. of reduce=, as long as the user
  knows what they are doing.

THE CURRENT UDR (USER-DEFINED REDUCTION OPERATION) INTERFACE - NOT FINAL

* A reduction class must implement the method

    accumulateOntoState(state, inputValue)

in addition to the other UDR requirements, to make reduce= available.

This statement in the above example:

    x reduce= some-computation;

translates into:

    globalOp.accumulateOntoState(x, some-computation);

where globalOp is an instance of the CoolReduction class
that is created upon entry of the forall loop.
At present the same globalOp is accessed by all iterations
of the forall loop. This may decrease performance due to
remote accesses if accumulateOntoState() accesses globalOp's fields.

* This operation, as well as the current UDR interface, are not final.

In the current implementation, storage for the accumulation state
is duplicated in the shadow variable and in the reduction class,
ex. instances of CoolReduction above. When used with a reduce intent,
that state is used only after the corresponding task completes.

We are planning to update our reduction interface to avoid such duplication.
We would also like to clarify and/or update the current accumulate()
and combine() operations.

Recall that currently, for a reduce intent, accumulate() is invoked
at the end of each task on:
* the task-local clone of globalOp (the receiver) and
* the shadow variable (the argument).
It adds the accumulation state from the shadow variable onto the clone op.
Then, combine() is invoked to add the state in the current task's op
onto its parent task's op - or the globalOp if there is no parent task.

HANDLING OF REDUCE= IN THE COMPILER

* Added reduce= token.

We might want a more general approach that allows op= for an arbitrary op.
I did not go for it for simplicity and because we do not have a firm plan.

Instead of a reduce= token, I could have the parser accept 'reduce'
followed by '='. This seems less ideal because it allows (requires?)
whitespace in between.

* Parser converts "x reduce= y" to:
  PRIM_REDUCE_ASSIGN(x,y)

* At start of normalization, PRIM_REDUCE_ASSIGN(x,y) is lowered to the call:
  globalOp.accumulateOntoState(x,y)

where globalOp comes from the reduce intent for x.

BLOCKSTMT::BYREFVARS

BlockStmt::byrefVars has changed its role since it was introduced,
without changing its name, to reduce revision diffs.
Currently it keeps a CallExpr-based representation of a with-clause
of a forall loop or a task construct. For a forall loop, it also
serves to indicate the BlockStmt with the loop body, which
is used to implement forall intents.
Going forward, I would like to switch to a dedicated representation
of a with clause instead of using a CallExpr. At which point
the field should be renamed, ex., to withClause.

OTHER CHANGES

* Added a check that there are no PRIM_REDUCE_ASSIGN calls in the tree
after resolution. Really, there shouldn't be any after normalization.
We do not already have a loop over all gCallExprs in the post-normalization
check, and I did not want to add one.

* While there, I removed the default value for the char* 'op' argument of
buildAssignment() - because that default was unused and, more importantly,
there is an assertion against that default in the function itself.

* Added accumulateOntoState() to our standard ReduceScanOp subclasses.

* Extended reduceIntents.rst with accumulateOntoState() and reduce=.

PARSER/LEXER-generated files for this change will be committed separately.