Add promoted fast follower support
In order for fast followers to trigger, chpl__staticFastFollowCheck and
chpl__dynamicFastFollowCheck must be implemented for the follower types. For
promoted expressions, the compiler creates an iteratorRecord follower, but
previously the fast follower checks weren't implemented for iterator records.

This adds support for building the checks in the compiler. Ideally we would
just have straight module code that looked something like:

```chapel
// _iteratorRecord has a field for each formal in the promoted function. e.g.
// `A + 2.0 * C` results in a record with fields for each array and the real
proc chpl__dynamicFastFollowCheck(x: _iteratorRecord, lead) {
  // tuple that only has elements for fields that require promotion
  var promotion_tup: recordToPromotionTuple(x);
  var cur_tup_idx = 1;
  for param fieldNum in 1..numFields(x.type) {
    var field = getField(x.type, fieldNum);
    if requiresPromotion(field.type, x) {
      promotion_tup(cur_tup_idx) = field;
      cur_tup_idx += 1;
    }
  }
  return chpl__dynamicFastFollowCheckZip(promotion_tup, lead);
}
```

but at resolution time the iterator record has no fields. They are only
populated during iterator lowering, so we have to create the checks in the
compiler.

At promotion wrapper creation time (during resolution) we build up functions
that look similar to example above. We stuff each field of an iterator record
that requires promotion into a tuple, and then call the zippered fast follower
check on the tuple. However, since the iterator record fields don't exist
during resolution, we use a new primitive that maps the formal of the original
iterator function to a field in the iterator record.

The new primitive (PRIM_ITERATOR_RECORD_FIELD_VALUE_BY_FORMAL) takes an
iterator record and the formal that will later be transformed into a field. The
formal is used to determine the type. During iterator inlining the iterator
records fields are finally added, one for each formal of the wrapper iterator.
At this time, we now replace our primitive with a PRIM_GET_MEMBER_VALUE and
replace the formal with the iterator record's field.

The end result is that we're able to use fast followers for promoted
expressions. This doubles the performance of the promoted stream variant,
making it on par with the global version.