Only build up one set of fast follower checks for a pair of ref/val functions
We build up separate promotion wrappers for both the ref and value function in
a ref/val pair. However, the fast follower checks for both functions are
identical, so if we build them up twice we'll have ambiguous functions. This
just updates resolution to only build up the fast follower checks for one of
them.