Add variations of stream that should _not_ trigger fast followers
Add versions of test_stream_is_parallel and test_whole_array_stream_is_parallel
that use different domains for some of the arrays to lock-in cases where
promoted fast followers are (correctly) not triggered.

This is just adds correctness tests to make sure that we're building and
forwarding the fast follower checks correctly for promoted expressions.