Fixed zip((...t)) case
It turns out that zippering over a tuple expansion was creating
problems because it has one argument (so was being treated like other
one-argument cases) which later became multiple arguments.  Here,
the fix is to just have the ForLoop rewriting pass remove the
tuple expansion, and to create a version of _getIterator() that
takes a tuple.  (Note that this is does not go against the goal
of making zippering not rely on tuples because it's OK to use
tuples if they already exist; simply not so cool to introduce new
tuples when they aren't necessary and limit our ability to iterate
over things like types).