Quick fix to allow new initializers to recursively call themselves
Due to the presence of the meme argument in initializers, we can give the
initializers a return type before trying to resolve its body, allowing
recursive calls that would otherwise fail (because we haven't and can't declare
our return type) to resolve and run correctly.  Old style constructors do not
have this argument and so cannot as simply benefit from this change (as their
_this field does not know its type until resolution of part of the body).

Ideally, this code will be removed when our compiler's handling of recursive
calls improves.