reduce intents: different input/state/output types
When specifying a reduce intent, allow specifying the input type
for the reduction operation. The output type is the type of the
outer variable. The shadow variable has the type of the accumulation
state, which is determined implicitly by the 'identity' method
of the ReduceScanOp subclass being used with the reduce intent.

See #4565 for detailed discussion.

The input type is specified as an argument to the reduction class
in the reduce intent, e.g.:

    forall ... with (MyReductionOp(int) reduce x) ....

I implemented this via explicit support in setupOneReduceIntent().
Avoid setting svar type explicitly in implementForallIntents.cpp
because in some cases it needs to be a value type.

Added tests.

Added an example to reduceIntents.rst.
While there, fixed an omission of 'ref' arg intent in reduceIntents.rst.