Re-enable anonymous for loop optimization for zippered loops
It turns out that the anonymous for loop optimization was brittle to
the structure of the zippered loop expressions coming down to it, and
had stopped firing for serial zippered loops with my recent changes.
Worse, because I've had to handle multiple cases in the compiler
rather than in the modules, and have some code in the old style and
other code in the new, the best way to keep it working seemed to push
the analysis into each branch of the rewrite I wrote (mixing it into
the building of the for loop rather than as a separate post-pass.
But it works.  Suggestions for improving it would be welcome.