Oops, fixing a crossed wire
In one of my previous commits, I renamed the routine to call
in the PRIM_EXPAND_TUPLE case _getIterator() where before it
had logically been _getIteratorZip().  Without realizing it,
this crossed two wires because it made the same routine get
called when iterating over a tuple as when trying to get an
iterator per tuple element.  This retires that overload and
makes the compiler call _getIteratorZip() instead.