Fixing two (2!) failures
Somehow, the only tests we have that exercise a zippered for expression
are the two knucleotide versions(?!) and somehow they found a flaw in
my code.  :'(

For now, I've reverted zippered for expressions back to old-style
tuple storage rather than PRIM_ZIPs which works around this issue.

This also motivated me to do a parser cleanup which occurred to me
after my original change, but which I'd been too lazy to do.
Essentially, I collapse my two par_zippered_iterator and
ser_zippered_iterator states back down to a single state that always
returns a PRIM_ZIP.  Then I convert it into a tuple in the cases that
can't yet handle a PRIM_ZIP on a case-by-case basis using a helper
function zipToTuple().  This has the benefit of (a) squashing some
essentially duplicate grammar rules down to one and (b) making it
clearer that it's not so much about serial and parallel zippering
(even though that's the line I was trying to draw) and more about the
new world vs. the old.