Range chunking library
Passes linux64 testing on chapcs. More testing and review to be done.

This is a library for dividing a bounded range of any stride into chunks and accessing them through an iterator or a zero-based index query. Code is shared between the two approaches with private helper methods.

The long term goal is to integrate this library into the Block distribution, but it is currently integrated with BlockCycDim, ReplicatedDim, DefaultSparse, and LayoutCSR.

This PR also updates tests that track standard modules, as RangeChunk is now one of them.