Improve QUICKSTART.rst and building instructions
This commit improves QUICKSTART.rst and the instructions
for building Chapel to be more succinct for running Chapel
on linux in common configurations. In particular, it takes
these steps:

 * present a combined command listing for bash
   (since bash is typically the default shell for linux)
 * move support for other shells to a later section
 * present command listings using `make` vs `gmake`
   since `make` is GNU make on every linux system I've
   seen (and many such linux systems don't have a `gmake`
   command or package).

Besides making the instructions more relevant for linux,
I was able to:

 * get the key starting information onto the first page or so of the
   QUICKSTART.rst instructions. This involved a few simple
   wording changes to shorted some sentences.

 * amend the QUICKSTART.rst instructions so that it's clearer
   which steps need to be taken if you already have Chapel
   installed (e.g. with Homebrew).

While there, I updated the links in QUICKSTART.rst to chapel.cray.com
to use 1.14. Since there is no 1.14 documentation currently at
chapel.cray.com, these links are currently broken.

As supporting changes, I updated building.rst to descibe how
to check to see if `make` is GNU make and to indicate that
parallel builds are supported. I also added a link target
to chplenv.rst for information about chplconfig files.