Update comm counts improved by my zip-without-tuples change
Yesterday's zip-without-tuples improved the comm counts for this
test -- something that hadn't even occurred to me could be possible.
Sorry not to have thought to have tested gasnet before merging...

Apparently, the lack of using tuples during iteration removes 4
on's (two fast, two not) for zippered serial iteration over remote
arrays.  To bad we don't do more of that in real codes... :)