New version of binary-trees
This is a proposed improvement to binary-trees for 1.14 in which
we use the new init() capability (and Lydia's fix for recursive
initializers) rather than a type method factory approach as before.
This simplifies the code and makes good use of the new init()
routines.

Updated the README to note the progress and add some new possible
TODOS.