Make use of new 'enum' features (.size / enumerator) in shootouts
This corresponds to commit 98f5a2a which I made a few weeks back
and makes use of the new 'enum' features in 1.14 in the release
versions of the benchmarks (thanks @Ldelaney!).  These had no
problems in the studies versions; probably I should've put them
into the release at that same time and been less cautious.