Add future for 'return' statements in initializers
While playing around with the binarytrees shootout, I made a
thinko in which I was doing 'return new C()' within an init()
routine for 'class C'.  This worked, but probably shouldn't
(an initializer shouldn't explicitly return anything, but
should initialize its 'this').  This future is designed to
capture this issue so that we can generate an error message
in the future.