Add some tests that demonstrate some current privatization bugs
Add some tests that directly insert and get objects from the runtime
privatization routines in order to show some current bugs and lock in the
correct behavior when they are fixed. Note that these tests are pretty specific
to the current runtime implementation and they will not be very portable if
that interface changes, but I think they are still incredibly valuable until
then, particularly given how subtle (and potentially bad) this bug is.

There are a couple of tests variations that:
 - add pids in order
 - add pids in parallel (random order)
 - add pids with large gaps between them
 - add pids while concurrently reading

All of these except for the pids in order fail spectacularly right now,
especially under valgrind.