Clean up the modules primer for better generated html
Fixed the rendering of the starting nested list, add back tics, reference to the
man page, and sections.  Also move some of the comments around so that the
generated code blocks look more intentional, and change some multiline //
comment to an actual multiline comment (leaving others that seem to be best
served as output comments).

Add a label to the man page so that it can be referenced from the modules primer
without errors.