Add sections to the ranges primer
@ben-albrecht suggested adding some section headers, so do that.
Add sections for "Ranges (general info)", "Operators", and "Slicing" and
for "Definition of writeRange" (just to get it out of the Slicing section).

Shift some of the example code around to put it into the right section. Update
a section header comment to talk about all operators instead of just 'by'
and 'align'

Update the .good to match the shifted code.