Add ability to skip multiple trials for long running tests
Update start_test so that we'll skip any remaining trials for tests that take
longer than CHPL_TEST_EXEC_TIME_LIMIT_NUM_TRIALS seconds to execute.

This is motivated by wanting to run a bunch of trials to gain confidence in the
release and to try to find any sporadic failures. It's not possible to run an
insane number of trials for long running tests, so we need a way to limit how
many trials we do for them.