Limit tests that don't work with multiple trials to 1 trial
Some tests do not work with multiple trials, so only allow them to do a single
trial. Most of these tests have prediffs that grep and then destroy generated
code or prediffs that examine and destroy some file/dirs for testing filesystem
routines.

Since the prediffs destroy the artifact being checked, the second trial fails,
so limit these test to 1 trial only.

This is motivated by wanting to run a bunch of trials to gain confidence in the
release and to try to find any sporadic failures. Since these tests fail with
more than one trial, it makes it hard to find any real failures, so just
limit them to one trial.