Updated distributions primer for online docs
* Markup.

* Some wordsmithing.

I backtick-ified only the first (one or a couple of) occurrence(s)
of each distribution name, treating the subsequent occurrences
as English names. My intention is to reduce the number of things
that stand out in the generated HTML and the amount of markup
for those reading the .chpl file directly.