Updated slices primer for online docs
In addition to markup, I switched all comments within main()
from //-style to /* */-style. This is so that those comments
render as plain text, rather than code. Other solutions:

* Get rid of main(). (This would be my personal preference.)
Then all //-style comments can start at the beginning of the line
and they will be rendered as plain text.

* Start //-style comments at the beginning of the line anyway,
ignoring indentation of the code.

The spacing of the /* */ comments is the best I came up with.
We can revisit if there are other suggestions.