Filter out JIRA-154 errors
We've been getting some sporadic gasnet failures along the lines of

    recursive failure in AMUDP_SPMDShutdown

There are several different error messages and sometimes it just shows up as
"binary files diff" because one of the errors has a null character in it.

These errors are ultimately caused by a halt (which leads to a gasnet_exit)
being called on a non-zero locale, leading to recursive shutdown calls, which
gasnet complains about. Greg and I are brainstorming solutions, but in the
meantime we need an easier way to identify the regression (previously there
were ~10 duplicate jira entires.)

This adds filtering for a handful of error messages, and adds some routines to
diff a binary file so we can see the error message for the error with a null
character.