In the technote on Atomics, add link to spec and atomics primer
When I was adding a spec reference to the atomics primer, I noticed that the
technote on Atomics didn't use a link to the spec either.  When I was fixing
that, I noticed that I could update a path to be a link to the atomics
primer.

Verified that the html output looked good and that the links work as expected.