Ensured consistent output for taskParallel.chpl
Note this is the test in parallel/taskPar/figueroa,
NOT release primers.

It was possible - however unlikely - that legal output of this test,
after being processed with .prediff, would fall short of .good.
There are two possible scenarios:

(1) In both invocations of player(), i.e. player(1) and player(2):

* The test notTired.readFF() in the while loop header succeeds
the first time around.

* The same test notTired.readFF() in the if statement fails,
because notTired has been set to false since it was tested
in the while loop.

When this happens, the output will miss the lines:

  Player {1,2}: {rock,paper,scissors}
  # or, after prediff-ing:
  Player X: rock, paper, or scissors

Note that the referee will still judge the game (call determineWinner())
because choice({1,2} gets set after each player() exists the while loop.
So the following line will still get printed:

  Referee: Player {1,2} wins!
  # or, after prediff-ing:
  Referee: Player X wins!

(2) By the time notTired.readFF() is checked in referee() in the while
loop the first time around, notTired has already been set to false.

In this case, both of the above line groups will be missing from the output.
This is what we probably observed in the whitebox nightly failure on 9/25

To ensure that both the players and the referee play at least one game,
and therefore to ensure matching of the prediff-ed output, I am convering
a while-do loop to do-while. I am also removing the duplicate check
for notTired.readFF() within the players' if statement.

The above scenarios can probably still happen, however that can be
only after each loop is executed once. Which should not affect the
output after prediff-ing, as it sorts out duplicate lines.