Remove the chplvis primers from the generated html due to redundancy
The chplvis tools description provides much of the same information as the
primers, including many of the code blocks.  It is also more extensive.  Make
it directly include the code it can include easily, and drop the primers from
their location within the primers hierarchy.

I updated the Makefile to copy the .chpl files to an expected location so that
it will be consistent, no matter the final layout of the test directory.  I
updated the primers index to no longer include the chplvis files.  I replaced
the code blocks for Examples 1, 2, and 4 with direct references to the copied
files (I did not do the same for Example 3 as it is more selective in the code
it demonstrates).  I updated the comment on Example 1's lead in to no longer
mention that the .chpl file has more comments (and had the include of that
file and the file for Example 2 ignore the comments referencing the tools
description's location).  I fixed a grammatical error, back tic'ed a reference
to "on statements", and updated the references in the primer .chpl files to
point to the current location of the chplvis.rst file.