Opaque domain/array primer review for online docs
* Markup, wordsmithing.

* Removed the comment about needing the module wrapper,
as the wrapper is not needed, as far as I can tell.
Todo post-release: remove the module wrapper.

After removing the module, we can also remove the line
with just "//" - that one serves to separate the
final "}" from the preceding code in the generated html.

* Consistently terminate sentences with a period.
I left unchanged those sentences that are
terminated with a colon.

* For the comments that are indented to go with
snippets of code, I left them indented as-is.
That way, they stand out less than the "top-level"
comments that start at the beginning of the line,
creating the "more important" vs. "less important"
comment hierarchy in the generated html.
This can be revisited if desired.