Fix for EXP1190 in GASNet testing
I was seeing problems with tests such as io/bradc/eof.chpl in
GASNet testing. During code generation, we'd be trying to
generate a PRIM_GET_SVEC_MEMBER to a field with reference type.
That generated a ref to a wide ref, which isn't what we want,
and confused code generation.

This commit adjusts preFold in function resolution to
change the PRIM_GET_SVEC_MEMBER in the homog. tuple accessor
to a PRIM_GET_SVEC_MEMBER_VALUE when the element is a reference
already.

Since a Chapel reference cannot be retargeted, there is no
real point in getting a reference to a reference. When setting
the reference field, the code should use
PRIM_SET_MEMBER_VALUE/PRIM_SET_SVEC_MEMBER.