Replace 4 FnSymbol fields + 1 FLAG with hashtable or, more precisely, with a std::map.
The need to do this was suggested by Mike in while reviewing #3838.

FLAG_PARTIAL_COPY and these fields of FnSymbol:

 partialCopyMap
 partialCopySource
 varargOldFormal
 varargNewFormals

were used internally during resolution. Our guideline for this case
is to use a hashtable instead of a field. This is what this change does.

DETAILS

* The existence of the entry partialCopyFnMap[fn] is equivalent to
the former fn->hasFlag(FLAG_PARTIAL_COPY).

* I made values in partialCopyFnMap be the actual data, rather than
a pointer to a struct/class allocated separately, to avoid indirection
and alloc/free. So there is an entry IFF there is data.
I was able to make that work nicely using helper functions.

I use a single hashtable for all four fields rather than a separate
hashtable for each field. This should reduce overhead e.g. hashing.

* I made partialCopyFnMap's keys to be node IDs. Since some FnSymbols
present in the map will be deleted from the tree and deallocated,
using FnSymbol* for their keys could result in bogus entries.

* Currently I am cleaning up partialCopyFnMap at the end of resolution
and verifying that it is empty at the end of each pass. My use of
node IDs for keys allows us to relax the "clear at the end of each pass"
rule in the future, should that be needed.

OPEN QUESTIONS

* How to switch partialCopyFnMap to be a hashtable?

When compiling hello4, it grows to 6100 entries, so using a binary tree
that std::map is is unfortunate. Ideally, make it std::unordered_map,
however that's not supported well by gcc 4.7. One way to address that is
a suggestion from David K:

#if __cplusplus < 201103
typedef std::map<a,b> partialCopyType
#else
typedef std::unordered_map<a,b> partialCopyType;
#endif

Or, resort to using our own HashMap from map.h.

* Should I put empty initializers for partialCopyMap and varargNewFormals
in the constructor for PartialCopyData?

* See if FnSymbol::retSymbol can be moved to PartialCopyData as well.
While there is a comment on the field saying "during partial copying,"
it appeared non-obvious that this is actually the case.